+++ packages/firebase/src/env.ts
@@ -1,15 +1,23 @@
 type Env = {
-  authenticationId: string;
-  apiKey: string;
+  storeId: number;
+  apiAuth: {
+    authenticationId: string;
+    apiKey: string;
+  },
   githubToken?: string;
 };
 
 export default () => {
   const {
+    ECOM_STORE_ID,
     ECOM_AUTHENTICATION_ID,
     ECOM_API_KEY,
     GITHUB_TOKEN,
   } = process.env;
+  const storeId = ECOM_STORE_ID && parseInt(ECOM_STORE_ID, 10);
+  if (!storeId) {
+    throw new Error('ECOM_STORE_ID is not set or not a number');
+  }
   if (!ECOM_AUTHENTICATION_ID) {
     throw new Error('ECOM_AUTHENTICATION_ID is not set');
   }
@@ -20,8 +28,11 @@ export default () => {
   const apiKey = ECOM_API_KEY;
   const githubToken = GITHUB_TOKEN;
   const env: Env = {
-    authenticationId,
-    apiKey,
+    storeId,
+    apiAuth: {
+      authenticationId,
+      apiKey,
+    },
     githubToken,
   };
   return env;