+++ packages/firebase/src/handlers/check-store-events.ts
@@ -1,12 +1,12 @@
 import type { EventSub } from '../types';
 // eslint-disable-next-line import/no-unresolved
 import { getFirestore } from 'firebase-admin/firestore';
-import { logger } from 'firebase-functions';
+import logger from 'firebase-functions/lib/logger';
 import api from '@cloudcommerce/api';
 import getEnv from '../env';
 
 export default async () => {
-  const { authenticationId, apiKey } = getEnv();
+  const { apiAuth } = getEnv();
   const eventsSubs = await getFirestore().collection('eventsSubs').get();
   const listenedEvents: EventSub['event'][] = [];
   eventsSubs.forEach((doc) => {
@@ -21,11 +21,8 @@ export default async () => {
     'products',
     'carts',
   ].forEach(async (resource) => {
-    const { data: { result } } = await api({
-      authenticationId,
-      apiKey,
-      endpoint: `events/${resource as 'orders'}`,
-    });
+    const { data: { result } } = await api
+      .get(`events/${resource as 'orders'}`, apiAuth);
     logger.info(`${resource} events: `, result);
   });
   return true;