+++ packages/firebase/src/index.ts
@@ -8,12 +8,24 @@ import { initializeApp } from 'firebase-admin/app';
 import functions from 'firebase-functions';
 import config from './config';
 import checkStoreEvents from './handlers/check-store-events';
+import handleAuthCallback from './handlers/auth-callback';
 
 initializeApp();
 
 const { httpsFunctionOptions: { region } } = config.get();
 
-export const cronStoreEvents = functions.region(region)
-  .pubsub.schedule('* * * * *').onRun(() => {
+const functionBuilder = functions
+  .region(region)
+  .runWith({
+    timeoutSeconds: 300,
+    memory: '128MB',
+  });
+
+export const cronStoreEvents = functionBuilder.pubsub
+  .schedule('* * * * *')
+  .onRun(() => {
     return checkStoreEvents();
   });
+
+export const appAuthCallback = functionBuilder.https
+  .onRequest(handleAuthCallback);